+++ src/hovers/hoverBase.ts
@@ -71,7 +71,8 @@ export abstract class HoverBase implements vscode.HoverProvider {
 
 	}
 
-	provideHover(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): vscode.ProviderResult<vscode.Hover> {
+	provideHover(document: vscode.TextDocument, position: vscode.Position): vscode.ProviderResult<vscode.Hover> {
+		// Prepare
 		const validchars="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789_";
 		let word='';
 		let p = position.character;
@@ -81,9 +82,17 @@ export abstract class HoverBase implements vscode.HoverProvider {
 		while (p > 0 && validchars.indexOf(line[p]) !== -1) { p--; }
 		// Skip leading invalid character
 		if (validchars.indexOf(line[p]) === -1) { p++; }
-		// Collect string until an invalid charecter is encountered
+		// Collect string until an invalid character is encountered
 		while (p < line.length && validchars.indexOf(line[p]) !== -1) { word += line[p++]; }
 
-		return new vscode.Hover(this.hoverText[word.toUpperCase()]);
+		// Found something to check for?
+		if (word) { 
+			// Search and validate
+			let content = this.hoverText[word.toUpperCase()];
+			if (content) { return new vscode.Hover(content); }
+		}
+
+		// Return
+		return undefined;
 	}
 }