+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The following enhancements and changes have been made to Atari Dev Studio:
 ## 0.5.8
 
 * Repackage to attempt fixing missing find-process package on macOS (Andrew Davie) 
+* Fixed issue where hover provider returned invalid content result (throwing unncessary errors in the developer logs)
 
 ## 0.5.7